+++ common/src/main/java/net/zelythia/TooltipHelper.java
@@ -16,13 +16,14 @@ import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentEffectComponents;
 import net.minecraft.world.item.enchantment.ItemEnchantments;
 import net.minecraft.world.item.enchantment.effects.EnchantmentValueEffect;
+import net.zelythia.config.AutoToolsConfig;
 
 import java.util.List;
 
 public class TooltipHelper {
 
     public static void applyTooltip(ItemStack stack, List<Component> tooltip) {
-        if (AutoToolsConfig.SHOWDPS) {
+        if (AutoToolsConfig.get().showDPS) {
             Item item = stack.getItem();
 
             if (item != Items.AIR) {