+++ CHANGELOG.md
@@ -1,1 +1,7 @@
+**Warning: This update changes the config file location and will reset your configurations**
+- Added config screen option to modify Silk Touch, Fortune and Shears block list to allow for further configuration
+- Custom tools now also support block and item tags + Better editing menu in config screen
+ 
+
 - Better compatability with mods that change enchantments like [Enchancement](https://modrinth.com/mod/enchancement)
+- Fixed a bug where shears would always be prioritized over tools with fortune or silk touch